feat(tools): graduate shared plumbing into tools/shared
The second workflow package is arriving (#334), which is the bar the
#335 amendment set for extraction. tools/shared now owns the
cross-workflow helpers: env contracts, the exec seam, the gh driver,
GitHub Actions plumbing (gha + gha.core), retry, and workspace
enumeration, with the @pnpm/* devDeps that back it. Per-file exports
(the release/workspace.ts pattern); consumers import by package name
with workspace:* deps.

- tools/release flips its ../lib imports to shared/…; src/lib is gone
- types.ts splits along the same seam: DependencyMap + PackageManifest
  (generic manifest vocabulary) graduate; DEP_FIELDS/DepField/Report
  stay with release's checks, which are their only consumers
- vue-check re-points to shared/workspace.ts; release drops its
  workspace.ts export — nothing consumes it anymore
- workspaceRoot loses one .. segment (tools/shared/src is three levels
  below the root); verified at runtime from both shared and vue-check
- pack.ts stays in release/src/checks per the amendment: publish-Pack
  semantics, check-only consumers
