+++ tools/release/src/checks/check-packed-manifest.ts
@@ -17,7 +17,7 @@ import {
   findPackedManifestViolations,
   formatPackedManifestReport,
 } from './check-pack.core.ts';
-import type { PackageManifest } from '../lib/types.ts';
+import type { PackageManifest } from 'shared/types.ts';
 
 const run = promisify(execFile);
