+++ tools/release/src/steps/release-tag.ts
@@ -8,8 +8,8 @@
 // tag fails here and that is fine. argv comes from the engine seam
 // (release-it.core.ts).
 
-import { boolEnv, requireEnv } from '../lib/env.core.ts';
-import { group, runMain } from '../lib/gha.ts';
+import { boolEnv, requireEnv } from 'shared/env.core.ts';
+import { group, runMain } from 'shared/gha.ts';
 import { tagArgs } from './release-it.core.ts';
 import { releaseItRun } from './release-it.ts';
