+++ tools/shared/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "shared",
+  "version": "0.0.1",
+  "private": true,
+  "description": "Cross-workflow tooling helpers: env contracts, exec/gh seams, GitHub Actions plumbing, retry, and workspace enumeration",
+  "type": "module",
+  "exports": {
+    "./env.core.ts": "./src/env.core.ts",
+    "./exec.ts": "./src/exec.ts",
+    "./gh.ts": "./src/gh.ts",
+    "./gha.core.ts": "./src/gha.core.ts",
+    "./gha.ts": "./src/gha.ts",
+    "./retry.ts": "./src/retry.ts",
+    "./types.ts": "./src/types.ts",
+    "./workspace.ts": "./src/workspace.ts"
+  },
+  "scripts": {
+    "format": "oxfmt \"**/*.{ts,js,json,jsonc,md}\"",
+    "format:check": "oxfmt --check \"**/*.{ts,js,json,jsonc,md}\"",
+    "lint": "oxlint .",
+    "test": "node --test \"src/**/*.test.ts\"",
+    "typecheck": "tsc -p tsconfig.json --noEmit"
+  },
+  "devDependencies": {
+    "@pnpm/fs.find-packages": "catalog:",
+    "@pnpm/workspace.read-manifest": "catalog:",
+    "@types/node": "catalog:",
+    "typescript": "catalog:"
+  }
+}
