+++ tools/shared/src/env.core.test.ts
