+++ tools/shared/src/env.core.ts
