+++ tools/shared/src/workspace.ts
@@ -18,10 +18,9 @@ import type { WorkspaceManifest } from '@pnpm/workspace.read-manifest';
 
 import type { PackageManifest } from './types.ts';
 
-/** Absolute path to the workspace root. This file lives in <root>/tools/release/src/lib. */
+/** Absolute path to the workspace root. This file lives in <root>/tools/shared/src. */
 export const workspaceRoot = join(
   dirname(fileURLToPath(import.meta.url)),
-  '..',
   '..',
   '..',
   '..',
