+++ tools/shared/turbo.json
@@ -1,0 +1,11 @@
+{
+  "$schema": "https://turborepo.com/schema.json",
+  "extends": ["//"],
+  "tasks": {
+    "test": {
+      // The root test task's src/** inputs miss the tsconfig/package surface
+      // these node:test suites depend on; over-approximate instead.
+      "inputs": ["$TURBO_DEFAULT$"]
+    }
+  }
+}
